Se resolvio la redundancia de las funciones de generar pin de transaccion
